+++ src/sail-codecs/fli/helpers.c
@@ -370,13 +370,13 @@ sail_status_t fli_private_encode_brun(struct sail_io* io, const unsigned char* p
 
 sail_status_t fli_private_decode_copy(struct sail_io* io, unsigned char* pixels, unsigned width, unsigned height)
 {
-    SAIL_TRY(io->strict_read(io->stream, pixels, width * height));
+    SAIL_TRY(io->strict_read(io->stream, pixels, (size_t)width * height));
     return SAIL_OK;
 }
 
 sail_status_t fli_private_encode_copy(struct sail_io* io, const unsigned char* pixels, unsigned width, unsigned height)
 {
-    SAIL_TRY(io->strict_write(io->stream, pixels, width * height));
+    SAIL_TRY(io->strict_write(io->stream, pixels, (size_t)width * height));
     return SAIL_OK;
 }
 
@@ -444,7 +444,7 @@ sail_status_t fli_private_decode_ss2(struct sail_io* io, unsigned char* pixels, 
     uint16_t lines_count;
     SAIL_TRY(io->strict_read(io->stream, &lines_count, sizeof(lines_count)));
 
-    for (uint16_t y = 0; y < lines_count && y < height; y++)
+    for (uint32_t y = 0; y < lines_count && y < height; y++)
     {
         unsigned char* line = pixels + y * width;
         uint16_t packets;
